requirements.txt
transformers
torch
pandas
pyttsx3
openai-whisper
sounddevice
numpy
